split dirs
