+++ ros/src/system/autoware_health_checker/include/health_aggregator/health_aggregator.h
