+++ ros/src/system/autoware_health_checker/include/node_status_publisher/rate_checker.h
