+++ ros/src/system/autoware_health_checker/include/node_status_publisher/value_manager.h
