+++ ros/src/system/autoware_health_checker/include/system_status_subscriber/system_status_subscriber.h
