+++ ros/src/system/autoware_health_checker/launch/health_checker.launch
@@ -3,6 +3,6 @@
     <param name="enable_statistics" value="true"/>
     <node pkg="autoware_health_checker" type="health_aggregator" name="health_aggregator" output="log" respawn="true" respawn_delay="0" />
     <node pkg="autoware_health_checker" type="health_analyzer" name="health_analyzer" output="log" respawn="true" respawn_delay="0">
-        <param name="warn_threashold" value="30"/>
+        <param name="warn_count_threshold" value="30"/>
     </node>
 </launch>
