+++ ros/src/system/autoware_health_checker/src/node_status_publisher/diag_buffer.cpp
@@ -17,7 +17,7 @@
  * v1.0 Masaya Kataoka
  */
 
-#include <autoware_health_checker/diag_buffer.h>
+#include <node_status_publisher/diag_buffer.h>
 
 namespace autoware_health_checker {
 DiagBuffer::DiagBuffer(std::string key, uint8_t type, std::string description,
