+++ ros/src/system/autoware_health_checker/src/node_status_publisher/rate_checker.cpp
@@ -17,7 +17,7 @@
  * v1.0 Masaya Kataoka
  */
 
-#include <autoware_health_checker/rate_checker.h>
+#include <node_status_publisher/rate_checker.h>
 
 namespace autoware_health_checker {
 RateChecker::RateChecker(double buffer_length, double warn_rate,
