+++ ros/src/system/autoware_health_checker/src/node_status_publisher/value_manager.cpp
@@ -17,7 +17,7 @@
  * v1.0 Masaya Kataoka
  */
 
-#include <autoware_health_checker/value_manager.h>
+#include <node_status_publisher/value_manager.h>
 
 ValueManager::ValueManager(ros::NodeHandle nh,ros::NodeHandle pnh)
 {
