+++ ros/src/system/autoware_health_checker/src/system_status_subscriber/system_status_subscriber.cpp
@@ -17,7 +17,7 @@
  * v1.0 Masaya Kataoka
  */
 
-#include <autoware_health_checker/system_status_subscriber.h>
+#include <system_status_subscriber/system_status_subscriber.h>
 
 namespace autoware_health_checker {
 SystemStatusSubscriber::SystemStatusSubscriber(ros::NodeHandle nh,
